// frontend/src/app/api/model/ingredientDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface IngredientDto { 
    id?: string;
    name: string;
    quantity: number;
    unit: string;
    expirationDate?: string;
    calories?: number;
    protein?: number;
    fat?: number;
    carbohydrates?: number;
}